+++ swim/macros/(Un-)Stun.js
@@ -1,3 +1,4 @@
+//!DOESN'T REROLL ON BENNY; USES UNSHAKE CODE INSTEAD!
 main();
 
 async function main() {