+++ swim/scripts/swimTab.js
@@ -1,0 +1,3 @@
+export class swimTab {
+    //miracle of a code here.
+}